+++ src/main/php/web/filters/BehindProxy.class.php
@@ -13,22 +13,37 @@ class BehindProxy implements Filter {
   private $remote, $replace;
 
   /**
-   * Creates a new instance given a map of the front-facing URL to the local path
+   * Creates a new instance given a map of the front-facing URL to the local path:
    *
-   * @param  [:string] $mapping e.g. `["http://remote.url/" => "/local.path"]`
+   * ```php
+   * // Using a map
+   * new BehindProxy(['http://remote.url/' => '/local.path']);
+   *
+   * // Specifying remote URL and local path as separate arguments
+   * new BehindProxy('http://remote.url/', '/local.path');
+   * ```
+   *
+   * @param  [:string]|string $mapping
+   * @param  ?string $target
    * @throws lang.IllegalArgumentException
    */
-  public function __construct($mapping) {
-    if (1 !== sizeof($mapping)) {
-      throw new IllegalArgumentException('Expected a map in the form ["http://remote.url/" => "/local.path"]');
+  public function __construct($mapping, $target= null) {
+    if (is_array($mapping)) {
+      if (1 !== sizeof($mapping)) {
+        throw new IllegalArgumentException('Expected a map in the form ["http://remote.url/" => "/local.path"]');
+      }
+
+      $this->remote= new URI(key($mapping));
+      $target= current($mapping);
+    } else {
+      $this->remote= new URI($mapping);
     }
 
-    $this->remote= new URI(key($mapping));
     if ($this->remote->isRelative()) {
       throw new IllegalArgumentException('Remote URL must be absolute');
     }
 
-    $this->replace= '#^'.rtrim(preg_quote(current($mapping), '#'), '/').'/#';
+    $this->replace= '#^'.rtrim(preg_quote($target, '#'), '/').'/#';
   }
 
   /**